Add right panel
