+++ crates/ruff_wasm/Cargo.toml
@@ -21,6 +21,7 @@ default = ["console_error_panic_hook"]
 ruff = { path = "../ruff" }
 ruff_diagnostics = { path = "../ruff_diagnostics" }
 ruff_python_ast = { path = "../ruff_python_ast" }
+ruff_python_formatter = { path = "../ruff_python_formatter" }
 ruff_rustpython = { path = "../ruff_rustpython" }
 
 console_error_panic_hook = { version = "0.1.7", optional = true }
@@ -30,7 +31,7 @@ rustpython-parser = { workspace = true }
 serde = { workspace = true }
 serde-wasm-bindgen = { version = "0.5.0" }
 wasm-bindgen = { version = "0.2.84" }
+js-sys = { version = "0.3.61" }
 
 [dev-dependencies]
-js-sys = { version = "0.3.61" }
 wasm-bindgen-test = { version = "0.3.34" }
