+++ playground/src/Editor/Editor.tsx
@@ -1,19 +1,26 @@
-import { useCallback, useEffect, useMemo, useState } from "react";
+import {
+  useCallback,
+  useDeferredValue,
+  useEffect,
+  useMemo,
+  useState,
+} from "react";
 import { DEFAULT_PYTHON_SOURCE } from "../constants";
-import init, {
-  check,
-  Diagnostic,
-  currentVersion,
-  defaultSettings,
-} from "../pkg";
+import init, { Diagnostic, Workspace } from "../pkg";
 import { ErrorMessage } from "./ErrorMessage";
 import Header from "./Header";
 import { useTheme } from "./theme";
 import { persist, restore, stringify } from "./settings";
 import SettingsEditor from "./SettingsEditor";
 import SourceEditor from "./SourceEditor";
-import MonacoThemes from "./MonacoThemes";
-import SideBar from "./SideBar";
+import { Panel, PanelGroup } from "react-resizable-panels";
+import PrimarySideBar from "./PrimarySideBar";
+import SecondarySideBar from "./SecondarySideBar";
+import { HorizontalResizeHandle } from "./ResizeHandle";
+import SecondaryPanel, {
+  SecondaryPanelResult,
+  SecondaryTool,
+} from "./SecondaryPanel";
 
 type Tab = "Source" | "Settings";
 
@@ -26,6 +33,7 @@ interface Source {
 interface CheckResult {
   diagnostics: Diagnostic[];
   error: string | null;
+  secondary: SecondaryPanelResult;
 }
 
 export default function Editor() {
@@ -33,6 +41,7 @@ export default function Editor() {
   const [checkResult, setCheckResult] = useState<CheckResult>({
     diagnostics: [],
     error: null,
+    secondary: null,
   });
   const [source, setSource] = useState<Source>({
     pythonSource: "",
@@ -42,15 +51,18 @@ export default function Editor() {
 
   const [tab, setTab] = useState<Tab>("Source");
   const [theme, setTheme] = useTheme();
+  const [secondaryTool, setSecondaryTool] = useState<SecondaryTool | null>(
+    null,
+  );
 
   const initialized = ruffVersion != null;
 
   useEffect(() => {
     init().then(() => {
-      setRuffVersion(currentVersion());
+      setRuffVersion(Workspace.version());
 
       const [settingsSource, pythonSource] = restore() ?? [
-        stringify(defaultSettings()),
+        stringify(Workspace.defaultSettings()),
         DEFAULT_PYTHON_SOURCE,
       ];
 
@@ -62,28 +74,72 @@ export default function Editor() {
     });
   }, []);
 
+  const deferredSource = useDeferredValue(source);
+
   useEffect(() => {
     if (!initialized) {
       return;
     }
 
-    const { settingsSource, pythonSource } = source;
+    const { settingsSource, pythonSource } = deferredSource;
 
     try {
       const config = JSON.parse(settingsSource);
-      const diagnostics = check(pythonSource, config);
+      const workspace = new Workspace(config);
+      const diagnostics = workspace.check(pythonSource);
+
+      let secondary: SecondaryPanelResult = null;
+
+      try {
+        switch (secondaryTool) {
+          case "AST":
+            secondary = {
+              status: "ok",
+              content: workspace.parse(pythonSource),
+            };
+            break;
+
+          case "Format":
+            secondary = {
+              status: "ok",
+              content: workspace.format(pythonSource),
+            };
+            break;
+
+          case "FIR":
+            secondary = {
+              status: "ok",
+              content: workspace.format_ir(pythonSource),
+            };
+            break;
+
+          case "Tokens":
+            secondary = {
+              status: "ok",
+              content: workspace.tokens(pythonSource),
+            };
+            break;
+        }
+      } catch (error: unknown) {
+        secondary = {
+          status: "error",
+          error: error instanceof Error ? error.message : error + "",
+        };
+      }
 
       setCheckResult({
         diagnostics,
         error: null,
+        secondary,
       });
     } catch (e) {
       setCheckResult({
         diagnostics: [],
         error: (e as Error).message,
+        secondary: null,
       });
     }
-  }, [initialized, source]);
+  }, [initialized, deferredSource, secondaryTool]);
 
   const handleShare = useMemo(() => {
     if (!initialized) {
@@ -111,6 +167,8 @@ export default function Editor() {
     }));
   }, []);
 
+  // useMonacoTheme();
+
   return (
     <main className="flex flex-col h-full bg-ayu-background dark:bg-ayu-background-dark">
       <Header
@@ -121,26 +179,51 @@ export default function Editor() {
         onShare={handleShare}
       />
 
-      <MonacoThemes />
-
       <div className="flex flex-grow">
         {initialized ? (
-          <>
-            <SideBar onSelectTool={(tool) => setTab(tool)} selected={tab} />
-            <SourceEditor
-              visible={tab === "Source"}
-              source={source.pythonSource}
-              theme={theme}
-              diagnostics={checkResult.diagnostics}
-              onChange={handlePythonSourceChange}
+          <PanelGroup direction="horizontal" autoSaveId="main">
+            <PrimarySideBar
+              onSelectTool={(tool) => setTab(tool)}
+              selected={tab}
             />
-            <SettingsEditor
-              visible={tab === "Settings"}
-              source={source.settingsSource}
-              theme={theme}
-              onChange={handleSettingsSourceChange}
+            <Panel id="main" order={0} className="my-2">
+              <SourceEditor
+                visible={tab === "Source"}
+                source={source.pythonSource}
+                theme={theme}
+                diagnostics={checkResult.diagnostics}
+                onChange={handlePythonSourceChange}
+              />
+              <SettingsEditor
+                visible={tab === "Settings"}
+                source={source.settingsSource}
+                theme={theme}
+                onChange={handleSettingsSourceChange}
+              />
+            </Panel>
+            {secondaryTool != null && (
+              <>
+                <HorizontalResizeHandle />
+                <Panel id="secondary-panel" order={1} className={"my-2"}>
+                  <SecondaryPanel
+                    theme={theme}
+                    tool={secondaryTool}
+                    result={checkResult.secondary}
+                  />
+                </Panel>
+              </>
+            )}
+            <SecondarySideBar
+              selected={secondaryTool}
+              onSelected={(tool) => {
+                if (secondaryTool === tool) {
+                  setSecondaryTool(null);
+                } else {
+                  setSecondaryTool(tool);
+                }
+              }}
             />
-          </>
+          </PanelGroup>
         ) : null}
       </div>
       {checkResult.error && tab === "Source" ? (
