+++ playground/src/Editor/SourceEditor.tsx
@@ -2,9 +2,9 @@
  * Editor for the Python source code.
  */
 
-import Editor, { useMonaco } from "@monaco-editor/react";
+import Editor, { BeforeMount, Monaco } from "@monaco-editor/react";
 import { MarkerSeverity, MarkerTag } from "monaco-editor";
-import { useCallback, useEffect } from "react";
+import { useCallback, useEffect, useRef } from "react";
 import { Diagnostic } from "../pkg";
 import { Theme } from "./theme";
 
@@ -21,7 +21,8 @@ export default function SourceEditor({
   theme: Theme;
   onChange: (pythonSource: string) => void;
 }) {
-  const monaco = useMonaco();
+  const monacoRef = useRef<Monaco | null>(null);
+  const monaco = monacoRef.current;
 
   useEffect(() => {
     const editor = monaco?.editor;
@@ -98,14 +99,21 @@ export default function SourceEditor({
     [onChange],
   );
 
+  const handleMount: BeforeMount = useCallback(
+    (instance) => (monacoRef.current = instance),
+    [],
+  );
+
   return (
     <Editor
+      beforeMount={handleMount}
       options={{
         readOnly: false,
         minimap: { enabled: false },
         fontSize: 14,
         roundedSelection: false,
         scrollBeyondLastLine: false,
+        contextmenu: false,
       }}
       language={"python"}
       wrapperProps={visible ? {} : { style: { display: "none" } }}
